refactor: todoStore 처리 함수 분리
